MBC: Re-write
- MBC (handler): re-wrote how the handling works, in preperation for supporting the other MBC models (2, 3, 5 etc).

- MBC1: Fixed a slew of bugs due to incorrect assumptions. More games now work.

+++ src/includes/rom.h
@@ -27,7 +27,7 @@ class Rom
 		static u8 mbcType;
 		static u8 romSize;
 		static u8 ramSize;
-		static u8 romBank;
+		static u16 romBank;
 		static u8 ramBank;
 		static u8 currentMode;
 		static bool hasBatteryBackup;
